requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.10 
djangorestframework==3.15.2
docopt==0.6.2
idna==3.10
Js2Py==0.74
numpy==1.24.4
packaging==24.2
pandas==2.2.3
pipwin==0.5.2
psycopg2>=2.9
psycopg2-binary==2.9.6
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
